+++ src/app/components/projects/pages/list/list.component.html
@@ -1,0 +1,12 @@
+<baw-model-list [modelKey]="'projects'" [filterPlaceholder]="'Filter projects'">
+  <!--
+    This error message is quite generic because it can appear under multiple
+    conditions:
+      1. There are no projects in the system
+      2. The user does not have permission to view any projects
+      3. The user filtered the list to something that has no results
+  -->
+  <ng-template #noResultsTemplate>
+    No projects found
+  </ng-template>
+</baw-model-list>
