+++ src/app/components/shared/model-list/model-list.component.html
@@ -23,7 +23,7 @@
     </ng-template>
   </li>
 
-  @if (siteConfig.settings.websiteMaps) {
+  @if (siteConfig.settings.homePageAndProjectListMaps) {
     <li [ngbNavItem]="tabs.map">
       <a ngbNavLink>
         <fa-icon [icon]="['fas', 'map-location-dot']"></fa-icon>
