+++ src/app/services/config/configMock.service.ts
@@ -60,7 +60,7 @@ export const testApiConfig = new Configuration({
       harvestFilenameGuide: "http://broken_link",
     },
     hideProjects: false,
-    websiteMaps: true,
+    homePageAndProjectListMaps: true,
     customMenu: [
       {
         title: "<< content1 >>",
