+++ src/assets/environment.json
@@ -26,7 +26,7 @@
       "harvestFilenameGuide": "https://github.com/ecoacoustics/metadata-standard/blob/master/filenames.md"
     },
     "hideProjects": false,
-    "websiteMaps": true,
+    "homePageAndProjectListMaps": true,
     "customMenu": [
       {
         "title": "Research",
